change to export parsed version
simplify api to avoid object creation. Allows a developer to just
require the module and get the parsed version object right away.
