+++ Readme.md
@@ -1,28 +1,28 @@
-
 # Node-version
-      
-  A quick module that returns current node version.
 
-  See server.js in examples/
+A quick module that returns current node version parsed into parts.
 
 ## Installation
 
-    npm install node-version
+```shell
+  npm install node-version
+```
 
 ## Quick Start
 
-    var nodeVersion = new (require('node-version').version);
-    var currentVersion = nodeVersion.getVersion();
+```js
+var currentVersion = require('node-version');
 
-    /*
-    console.log(currentVersion);
+/*
+console.log(currentVersion);
 
-    {
-    	original: 'v0.4.10', // same as process.version
-  		short: '0.4',
-  		long: '0.4.10',
-  		major: '0',
-  		minor: '4',
-  		build: '10'
-  	}
-    */
+{
+    original: 'v0.4.10', // same as process.version
+    short: '0.4',
+    long: '0.4.10',
+    major: '0',
+    minor: '4',
+    build: '10'
+}
+*/
+```
